+++ CHANGES.md
@@ -1,3 +1,6 @@
+# Oct 30, 2017
+    - Audit login and logout
+
 # Oct 10, 2017
     - Implement social media integration